NavContainer refactor
