+++ aries-core/src/js/components/core/NavigationMenu/NavContainer/NavHeader.tsx
@@ -1,6 +1,6 @@
 import { Box, Button, Heading } from 'grommet';
 import { Sidebar as SidebarIcon } from 'grommet-icons';
-import { ScreenReaderOnly } from '../ScreenReaderOnly';
+import { ScreenReaderOnly } from '../../ScreenReaderOnly';
 
 interface NavHeaderProps {
   title?: string;
@@ -8,7 +8,7 @@ interface NavHeaderProps {
   setOpen: (open: boolean) => void;
 }
 
-const NavHeader = ({ title, open, setOpen, ...rest }: NavHeaderProps) => {
+export const NavHeader = ({ title, open, setOpen, ...rest }: NavHeaderProps) => {
   const heading = (
     <Heading level={2} size="small" margin="none">
       {title}
@@ -42,34 +42,5 @@ const NavHeader = ({ title, open, setOpen, ...rest }: NavHeaderProps) => {
         />
       </Box>
     </>
-  );
-};
-
-interface NavContainerProps {
-  children: React.ReactNode;
-  header?: React.ReactNode;
-  open: boolean;
-  setOpen: (open: boolean) => void;
-  title?: string;
-  [key: string]: any; // Allow other props like pad, width, etc.
-}
-
-export const NavContainer = ({
-  children,
-  header,
-  open,
-  setOpen,
-  title,
-  ...rest
-}: NavContainerProps) => {
-  return (
-    <Box
-      pad={{ horizontal: 'small' }}
-      width={open ? { min: 'small' } : undefined}
-      {...rest}
-    >
-      {header || <NavHeader title={title} open={open} setOpen={setOpen} />}
-      {children}
-    </Box>
   );
 };
