+++ aries-core/src/js/components/core/NavigationMenu/NavContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './NavContainer';
